Fully move notes to notes module
