--- app/src/main/java/com/bhaktaprogram/notehelper/features/notes/domain/NoteInfo.kt
@@ -1,6 +1,0 @@
-package com.bhaktaprogram.notehelper.features.notes.domain
-
-data class NoteInfo(
-    val id: Int,
-    val title: String
-)
