--- app/src/main/java/com/bhaktaprogram/notehelper/features/notes/domain/NotesRepository.kt
@@ -1,6 +1,0 @@
-package com.bhaktaprogram.notehelper.features.notes.domain
-
-interface NotesRepository {
-
-    suspend fun getAll(): List<NoteInfo>
-}
