--- app/src/main/java/com/bhaktaprogram/notehelper/features/notes/repository/NotesDataSource.kt
@@ -1,8 +1,0 @@
-package com.bhaktaprogram.notehelper.features.notes.repository
-
-import com.bhaktaprogram.notehelper.features.notes.domain.NoteInfo
-
-interface NotesDataSource {
-
-    suspend fun getAll(): List<NoteInfo>
-}
