--- app/src/main/java/com/bhaktaprogram/notehelper/features/notes/ui/NotesViewModelFactory.kt
@@ -1,16 +1,0 @@
-package com.bhaktaprogram.notehelper.features.notes.ui
-
-import androidx.lifecycle.ViewModel
-import androidx.lifecycle.ViewModelProvider
-import javax.inject.Inject
-import javax.inject.Provider
-
-class NotesViewModelFactory @Inject constructor(
-    private val provider: Provider<NotesViewModel>
-) : ViewModelProvider.Factory {
-
-    @Suppress("UNCHECKED_CAST")
-    override fun <T : ViewModel?> create(modelClass: Class<T>): T {
-        return provider.get() as T
-    }
-}
