+++ features/notes/build.gradle
@@ -33,6 +33,9 @@ android {
     kotlinOptions {
         jvmTarget = versions.java
     }
+    buildFeatures {
+        viewBinding true
+    }
 }
 
 dependencies {
