+++ features/notes/src/main/java/com/bhaktaprogram/notehelper/notes/domain/NotesInteractor.kt
@@ -2,9 +2,10 @@ package com.bhaktaprogram.notehelper.notes.domain
 
 import javax.inject.Inject
 
-internal class NotesInteractor @Inject constructor(
+class NotesInteractor @Inject constructor(
     private val repository: NotesRepository
 ) {
+    suspend fun getAll(): List<NoteInfo> = repository.getAll()
 
 
 }
