+++ features/notes/src/main/java/com/bhaktaprogram/notehelper/notes/ui/NotesFragment.kt
@@ -1,32 +1,51 @@
 package com.bhaktaprogram.notehelper.notes.ui
 
 import android.os.Bundle
-import android.view.LayoutInflater
 import android.view.View
-import android.view.ViewGroup
 import androidx.fragment.app.Fragment
 import androidx.lifecycle.ViewModelProvider
+import androidx.lifecycle.lifecycleScope
 import com.bhaktaprogram.notehelper.notes.R
+import com.bhaktaprogram.notehelper.notes.databinding.NotesFragmentBinding
+import com.bhaktaprogram.notehelper.notes.di.DaggerNotesComponent
+import com.bhaktaprogram.notehelper.notes.domain.NoteInfo
+import kotlinx.coroutines.flow.collect
+import javax.inject.Inject
 
-class NotesFragment : Fragment() {
+class NotesFragment : Fragment(R.layout.notes_fragment) {
 
-    companion object {
-        fun newInstance() = NotesFragment()
-    }
-
+    @Inject
+    lateinit var viewModelFactory: ViewModelProvider.Factory
     private lateinit var viewModel: NotesViewModel
+    private var _binding: NotesFragmentBinding? = null
+    private val binding get() = _binding!!
 
-    override fun onCreateView(
-        inflater: LayoutInflater, container: ViewGroup?,
-        savedInstanceState: Bundle?
-    ): View? {
-        return inflater.inflate(R.layout.notes_fragment, container, false)
+    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
+        super.onViewCreated(view, savedInstanceState)
+
+        DaggerNotesComponent.create().inject(this)
+
+        _binding = NotesFragmentBinding.bind(view)
+
+        viewModel = ViewModelProvider(this, viewModelFactory)
+            .get(NotesViewModel::class.java)
+
+        lifecycleScope.launchWhenStarted {
+            viewModel.state.collect(::updateState)
+        }
     }
 
-    override fun onActivityCreated(savedInstanceState: Bundle?) {
-        super.onActivityCreated(savedInstanceState)
-        viewModel = ViewModelProvider(this).get(NotesViewModel::class.java)
-        // TODO: Use the ViewModel
+    private fun updateState(list: List<NoteInfo>) {
+        val notes = StringBuilder("Notes:\n")
+        list.forEach {
+            notes.append("\t").append(it.toString()).append("\n")
+        }
+        notes.trim()
+        binding.textView.text = notes
     }
 
+    override fun onDestroyView() {
+        super.onDestroyView()
+        _binding = null
+    }
 }
